requirements.txt
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Deprecated==1.3.1
ecdsa==0.19.1
Flask==3.1.2
Flask-Caching==2.3.1
Flask-Limiter==3.11.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
mysql-connector-python==9.4.0
ordered-set==4.1.0
packaging==24.2
pyasn1==0.6.1
Pygments==2.19.2
python-jose==3.5.0
rich==13.9.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wrapt==2.0.1
zipp==3.23.0
